+++ digitama/sprite/ulpc.cpp
@@ -45,23 +45,23 @@ static inline const char* random_elf_name(int hint) {
 /*************************************************************************************************/
 WarGrey::AoC::SpriteULPCSheet::SpriteULPCSheet(const std::string& pathname) : SpriteGridSheet(pathname, 21, 13) {}
 
-void WarGrey::AoC::SpriteULPCSheet::on_heading_changed(float theta_rad, float vx, float vy, float prev_vr) {
+void WarGrey::AoC::SpriteULPCSheet::on_heading_changed(double theta_rad, double vx, double vy, double prev_vr) {
     I4WayMotion::dispatch_heading_event(theta_rad, vx, vy, prev_vr);
 }
 
-void WarGrey::AoC::SpriteULPCSheet::on_eward(float theta_rad, float vx, float vy) {
+void WarGrey::AoC::SpriteULPCSheet::on_eward(double theta_rad, double vx, double vy) {
     this->play("rwalk");
 }
 
-void WarGrey::AoC::SpriteULPCSheet::on_wward(float theta_rad, float vx, float vy) {
+void WarGrey::AoC::SpriteULPCSheet::on_wward(double theta_rad, double vx, double vy) {
     this->play("lwalk");
 }
 
-void WarGrey::AoC::SpriteULPCSheet::on_sward(float theta_rad, float vx, float vy) {
+void WarGrey::AoC::SpriteULPCSheet::on_sward(double theta_rad, double vx, double vy) {
     this->play("dwalk");
 }
 
-void WarGrey::AoC::SpriteULPCSheet::on_nward(float theta_rad, float vx, float vy) {
+void WarGrey::AoC::SpriteULPCSheet::on_nward(double theta_rad, double vx, double vy) {
     this->play("uwalk");
 }
